clean GPU code warnings

+++ lib/processes/eigenVis.cpp
@@ -23,12 +23,12 @@ eigenVis::eigenVis(Config& config,
     output_buffer = get_buffer("out_buf");
     register_producer(output_buffer, unique_name.c_str());
     num_eigenvectors =  config.get<uint32_t>(unique_name, "num_ev");
-    num_diagonals_filled =  config.get_default<int32_t>(unique_name,
+    num_diagonals_filled =  config.get_default<uint32_t>(unique_name,
                                                    "num_diagonals_filled", 0);
     // Read a list from the config, but permit it to be absent (implying empty).
     try {
-        for (int32_t e : config.get<std::vector<int32_t>>(unique_name,
-                                                          "exclude_inputs")) {
+        for (uint32_t e : config.get<std::vector<uint32_t>>(unique_name,
+                                                            "exclude_inputs")) {
             exclude_inputs.push_back(e);
         }
     } catch (std::runtime_error const & ex) {
@@ -44,7 +44,7 @@ void eigenVis::main_thread() {
 
     unsigned int input_frame_id = 0;
     unsigned int output_frame_id = 0;
-    unsigned int num_elements;
+    uint32_t num_elements;
     bool initialized = false;
     size_t lapack_failure_total = 0;
 
@@ -105,8 +105,8 @@ void eigenVis::main_thread() {
         // Fill the upper half (lower in fortran order!) of the square version
         // of the visibilities.
         int prod_ind = 0;
-        for (int i = 0; i < num_elements; i++) {
-            for (int j = i; j < i + num_diagonals_filled && j < num_elements; j++) {
+        for (uint32_t i = 0; i < num_elements; i++) {
+            for (uint64_t j = i; j < i + num_diagonals_filled && j < num_elements; j++) {
                 cfloat value = 0;
                 for (uint32_t ev_ind = 0; ev_ind < num_eigenvectors; ev_ind++) {
                     value += (std::conj(last_evs[freq_id][ev_ind * num_elements + i])
@@ -115,7 +115,7 @@ void eigenVis::main_thread() {
                 vis_square[i * num_elements + j] = value;
                 prod_ind++;
             }
-            for (int j = i + num_diagonals_filled; j < num_elements; j++) {
+            for (uint64_t j = i + num_diagonals_filled; j < num_elements; j++) {
                 // Conjugate because Fortran interprets as lower triangle.
                 vis_square[i * num_elements + j] = std::conj(input_frame.vis[prod_ind]);
                 prod_ind++;
@@ -124,10 +124,10 @@ void eigenVis::main_thread() {
 
         // Go through and zero out data in excluded rows and columns.
         for (auto iexclude : exclude_inputs) {
-            for (int j = 0; j < iexclude; j++) {
+            for (uint32_t j = 0; j < iexclude; j++) {
                 vis_square[j * num_elements + iexclude] = {0, 0};
             }
-            for (int j = iexclude; j < num_elements; j++) {
+            for (uint32_t j = iexclude; j < num_elements; j++) {
                 vis_square[iexclude * num_elements + j] = {0, 0};
             }
         }
@@ -160,7 +160,7 @@ void eigenVis::main_thread() {
 
         // Update the stored eigenvectors for the next iteration.
         for (uint32_t ev_ind = 0; ev_ind < num_eigenvectors; ev_ind++) {
-            for (int i = 0; i < num_elements; i++) {
+            for (uint32_t i = 0; i < num_elements; i++) {
                 last_evs[freq_id][ev_ind * num_elements + i] =
                     std::sqrt(evals[ev_ind]) * evecs[ev_ind * num_elements + i];
             }
@@ -244,7 +244,7 @@ void eigenVis::main_thread() {
             int indr = num_eigenvectors - 1 - i;
             output_frame.eval[i] = evals[indr];
 
-            for(int j = 0; j < num_elements; j++) {
+            for(uint32_t j = 0; j < num_elements; j++) {
                 output_frame.evec[i * num_elements + j] = evecs[indr * num_elements + j];
             }
         }
